requirements.txt
psycopg[c]===3.1.10
lxml===4.9.3
strawberry-graphql[fastapi]===0.211.0
fastapi===0.104.0
uvicorn===0.23.2